+++ mainwindow.h
@@ -31,12 +31,12 @@ private:
     Stylesheet stylesheet;
     Difficulty::DifficultyStruct difficulty;
 
-    qint32 buttonSize;
-    qint32 minesLeftFrameWidth;
-    qint32 minesLeftFrameHeight;
-    qint32 timerFrameHeight;
-    qint32 minesLeftNumberWidth;
-    qint32 spacerMiddleLeftFixedWidth;
+    quint16 buttonSize;
+    quint16 minesLeftFrameWidth;
+    quint16 minesLeftFrameHeight;
+    quint16 timerFrameHeight;
+    quint16 minesLeftNumberWidth;
+    quint16 spacerMiddleLeftFixedWidth;
 
     void clearLayout(QLayout*);
     void newGame(const Difficulty::DifficultyStruct&);
@@ -47,7 +47,7 @@ public:
     MainWindow(QWidget *parent = nullptr);
     ~MainWindow();
 
-    void setInfoBarNumber(QWidget*, const qint32&);
+    void setInfoBarNumber(QWidget*, const quint16&);
 private slots:
     void on_actionNew_triggered();
     void on_actionQuit_triggered();
@@ -58,7 +58,7 @@ private slots:
     void minesleft_changed_slot(const qint16&);
     void start_timer_slot();
     void solver_stopped_slot(const char&);
-    void set_infobar_time_slot(const QString&, const qint32&);
+    void set_infobar_time_slot(const QString&, const quint16&);
 protected:
     bool eventFilter(QObject*, QEvent*);
 };
